Cl2: Make CL2_* envs available in templates
This will make all env variables starting with CL2_ prefix available in the CL2 template files.

It will allow us to leverage the presets mechanism in cl2 templates and use it instead of relying on cumbersome test override files (in some cases).

The problem it solves is that sometimes we want to set something for almost all (but not all tests) tests.
Currently, the only way to do that is the create an override file and pass it to many test definitions.
This is cumbersome for many reasons, e.g. it requires at least 2 PRs (one to perf-test with the override file, second to test-infra).